+++ Example/KFXLocation/DEMOAppDelegate.m
@@ -1,10 +1,4 @@
-//
-//  DEMOAppDelegate.m
-//  KFXLocation
-//
-//  Created by ChristianFox on 07/05/2017.
-//  Copyright (c) 2017 ChristianFox. All rights reserved.
-//
+
 
 #import "DEMOAppDelegate.h"