+++ Example/KFXLocation/DEMOViewController.m
@@ -1,12 +1,18 @@
-//
-//  DEMOViewController.m
-//  KFXLocation
-//
-//  Created by ChristianFox on 07/05/2017.
-//  Copyright (c) 2017 ChristianFox. All rights reserved.
-//
+
+/********************************
+ *
+ * Copyright © 2016-2018 Christian Fox
+ *
+ * MIT Licence - Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
+ *
+ * This file is included with KFXLocation
+ *
+ ************************************/
+
+
 
 #import "DEMOViewController.h"
+#import <KFXLocation/KFXLocation.h>
 
 @interface DEMOViewController ()
 
@@ -14,16 +20,54 @@
 
 @implementation DEMOViewController
 
-- (void)viewDidLoad
-{
-    [super viewDidLoad];
-	// Do any additional setup after loading the view, typically from a nib.
+
+-(IBAction)openAppleMapsApp1ButtonTapped:(id)sender{
+    
+    KFXGeoLocationHelper *helper = [KFXGeoLocationHelper geoLocationHelper];
+    CLLocation *location = [[CLLocation alloc]initWithLatitude:51.0 longitude:-2.0];
+    [helper openAppleMapsAppWithDirectionsToLocation:location directionsMode:MKLaunchOptionsDirectionsModeTransit completionBlock:^(BOOL success, NSError * _Nullable error) {
+        
+        if (error != nil) {
+            NSLog(@"ERROR: %@",error);
+        }
+        
+    }];
 }
 
-- (void)didReceiveMemoryWarning
-{
-    [super didReceiveMemoryWarning];
-    // Dispose of any resources that can be recreated.
+-(IBAction)openAppleMapsApp2ButtonTapped:(id)sender{
+    
+    KFXGeoLocationHelper *helper = [KFXGeoLocationHelper geoLocationHelper];
+    CLLocation *location = [[CLLocation alloc]initWithLatitude:51.0 longitude:-2.0];
+    NSDictionary *options = @{
+                              MKLaunchOptionsDirectionsModeKey:MKLaunchOptionsDirectionsModeDriving,
+                              MKLaunchOptionsMapTypeKey:@3,
+                              MKLaunchOptionsShowsTrafficKey:@YES
+                              };
+    [helper openAppleMapsAppWithDirectionsToLocation:location
+                               mapLaunchOptions:options
+                                completionBlock:^(BOOL success, NSError * _Nullable error) {
+        
+        if (error != nil) {
+            NSLog(@"ERROR: %@",error);
+        }
+        
+    }];
 }
+
+
+- (IBAction)openGoogleMapsApp1ButtonTapped:(id)sender {
+    
+    KFXGeoLocationHelper *helper = [KFXGeoLocationHelper geoLocationHelper];
+    CLLocation *location = [[CLLocation alloc]initWithLatitude:51.0 longitude:-2.0];
+    [helper openGoogleMapsAppWithDirectionsToLocation:location
+                                       directionsMode:@"driving"
+                                      completionBlock:^(BOOL boolValue) {
+                                          NSLog(@"Success %d",boolValue);
+                                      }];
+
+}
+
+
+
 
 @end