+++ Example/Pods/KFXAdditions/KFXAdditions/Classes/Foundation/NSURL/NSURL+KFXAdditions.h
@@ -1,8 +1,8 @@
 /********************************
  *
- * Copyright © 2016-2017 Christian Fox
- * All Rights Reserved
- * Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
+ * Copyright © 2016-2018 Christian Fox
+ *
+ * MIT Licence - Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
  *
  * This file is included with KFXAdditions
  *
@@ -13,7 +13,7 @@
 
 @interface NSURL (KFXAdditions)
 
-
+/// Returns the keys and values of the parameters from the url
 -(NSDictionary<NSString*,NSString*>*)kfx_queryParameters;
 
 @end