+++ Example/Pods/KFXAdditions/KFXAdditions/Classes/Foundation/Strings/NSString+KFXAdditions.m
@@ -1,8 +1,8 @@
 /********************************
  *
- * Copyright © 2016-2017 Christian Fox
- * All Rights Reserved
- * Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
+ * Copyright © 2016-2018 Christian Fox
+ *
+ * MIT Licence - Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
  *
  * This file is included with KFXAdditions
  *
@@ -175,6 +175,43 @@
     }
     return finalIndex;
 }
+
+-(CGFloat)kfx_matchPercentageWithOtherString:(NSString *)visitorString{
+    
+    if ([self isEqualToString:visitorString]) {
+        return 1.0;
+    }
+    
+    NSInteger matchCount = 0;
+    NSInteger receiverLength = self.length;
+    NSInteger visitorLength = visitorString.length;
+    NSInteger longestLength;
+
+    if (receiverLength >= visitorLength) {
+        longestLength = receiverLength;
+        for (NSInteger idx = 0; idx < receiverLength; idx++) {
+            
+            if (idx < visitorLength) {
+                unichar rC = [self characterAtIndex:idx];
+                unichar vC = [visitorString characterAtIndex:idx];
+                matchCount += (rC == vC);
+            }
+        }
+    }else{
+        longestLength = visitorLength;
+        for (NSInteger idx = 0; idx < visitorLength; idx++) {
+            
+            if (idx < receiverLength) {
+                unichar rC = [self characterAtIndex:idx];
+                unichar vC = [visitorString characterAtIndex:idx];
+                matchCount += (rC == vC);
+            }
+        }
+    }
+    
+    return (CGFloat)matchCount / (CGFloat)longestLength;
+}
+
 
 //--------------------------------------------------------
 #pragma mark - New String with edits