+++ KFXLocation/Classes/KFXLocation.h
@@ -1,0 +1,19 @@
+/********************************
+ *
+ * Copyright © 2016-2018 Christian Fox
+ *
+ * MIT Licence - Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
+ *
+ * This file is included with KFXLocation
+ *
+ ************************************/
+
+
+
+#ifndef KFXLocation_h
+#define KFXLocation_h
+
+#import <KFXLocation/KFXLocationTracker.h>
+#import <KFXLocation/KFXGeoLocationHelper.h>
+
+#endif /* KFXLocation_h */