+++ KFXLocation/Classes/KFXLocationTracker.h
@@ -1,12 +1,14 @@
 /********************************
  *
- * Copyright © 2016-2017 Christian Fox
- * All Rights Reserved
- * Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
+ * Copyright © 2016-2018 Christian Fox
+ *
+ * MIT Licence - Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
  *
  * This file is included with KFXLocation
  *
  ************************************/
+
+
 
 // # Imports
 // Cocoa Frameworks
@@ -35,7 +37,7 @@ typedef void(^KFXLocationUpdatesBlock)(NSArray<CLLocation*> *locations);
 -(void)locationTracker:(KFXLocationTracker*)locationTracker didRejectLocation:(CLLocation*)location forReason:(KFXLocationRejectionReason)reason;
 -(void)locationTrackerDidPauseLocationUpdates:(KFXLocationTracker*)tracker;
 -(void)locationTrackerDidResumeLocationUpdates:(KFXLocationTracker*)tracker;
--(void)locationTracker:(KFXLocationTracker*)locationTracker didChangeAuthrizationStatus:(CLAuthorizationStatus)status;
+-(void)locationTracker:(KFXLocationTracker*)locationTracker didChangeAuthorizationStatus:(CLAuthorizationStatus)status;
 
 @end