+++ Example/CHANGELOG.md
@@ -1,6 +1,15 @@
 
 # ChangeLog - KFXLocation
 
+### v0.5.0 | Updates Licence, moves project to github, makes pod public
+
+### v0.4.2
+- Adds KFXLocation header
+- Adds methods to KFXGeoLocationHelper that opens the Apple Maps app
+- Adds methods to KFXGeoLocationHelper that opens the Google Maps app
+- Fixes a typo -locationTracker:didChangeAuthrizationStatus to -locationTracker:didChangeAuthorizationStatus
+- Now if -requestLocationTrackingAuthorisationWithBlock: is called but no location usage description is given in the info.plit the current status will be provided in the completion block.
+
 ### v0.4.1 - updates pods
 
 ### v0.4.0 - Updates pod dependencies - gets rid of KFXUtilities and adds KFXCore